Added audio and video files support

+++ bot.py
@@ -3,9 +3,11 @@ from ragSearch import RagSearch
 import telebot
 from telebot.types import InlineKeyboardMarkup, InlineKeyboardButton
 from tokens import TOKEN
+from download import is_download_link, get_file_name
 import os
+import requests
 
-model = Model("bge-m3:567m", "gemma3:4b")
+model = Model("bge-m3:567m", "gemma3:12b")
 ragDB = {}
 
 bot = telebot.TeleBot(TOKEN)
@@ -20,7 +22,7 @@ start_text = """Привет 👋
 Я — AI-ассистент, использующий RAG (Retrieval-Augmented Generation) поиск.
 
 Я умею:
-📄 Читать и анализировать файлы (PDF или txt)
+📄 Читать и анализировать файлы (pdf, txt, mp3, mp4)
 💬 Отвечать на вопросы по содержимому документов
 🔍 Искать нужную информацию по смыслу, а не по ключевым словам"""
 
@@ -37,10 +39,14 @@ def start(message):
 @bot.callback_query_handler(func=lambda x: x.data == "question")
 def question(call):
     message = call.message
-
     bot.edit_message_reply_markup(message.chat.id, prev_messages[message.chat.id].message_id, reply_markup=None)
-    bot.send_message(message.chat.id, "Напишите свой вопрос")
-    bot.register_next_step_handler(message, question_text)
+    
+    if len(ragDB[message.chat.id].get_files())==0:
+        m = bot.send_message(message.chat.id, "Сначала загрузите источники", reply_markup=keyboard)
+        prev_messages[message.chat.id] = m
+    else:       
+        bot.send_message(message.chat.id, "Напишите свой вопрос")
+        bot.register_next_step_handler(message, question_text)
         
 def question_text(message):
     sent = bot.send_message(message.chat.id, "Обработка вопроса...")
@@ -50,36 +56,74 @@ def question_text(message):
     prev_messages[message.chat.id] = m    
 
 
+def is_valid_attachment(message):
+    size = 52428800 # 50MB
+    return message.audio and message.audio.file_size<size or \
+           message.document and message.document.file_size<size or \
+           message.video and message.video.file_size<size
+       
+
+
 
 @bot.callback_query_handler(func=lambda x: x.data == "file")
 def add_file(call):
     message = call.message
 
     bot.edit_message_reply_markup(message.chat.id, prev_messages[message.chat.id].message_id, reply_markup=None)
-    bot.send_message(message.chat.id, "Пришлите источник в виде файла .pdf или .txt")
+    bot.send_message(message.chat.id, "Пришлите источник в виде файла .pdf, .txt, mp3 или mp4 размером до 50MB. Так же можно прислать ссылку на скачивание")
     bot.register_next_step_handler(message, file)
 
 def file(message):
-    sent = bot.send_message(message.chat.id, "Обработка файла...")
+    sent = bot.send_message(message.chat.id, "Cкачивание файла...")
 
-    file_info = bot.get_file(message.document.file_id)
-    downloaded_file = bot.download_file(file_info.file_path)
+    if (message.text and is_download_link(message.text)):
+        response = requests.get(message.text)
+        if response.status_code == 200:
+            downloaded_file = response.content
+            file_name = get_file_name(message.text)
+            if not file_name:
+                downloaded_file = None
+        else:
+            downloaded_file = None
 
-    file_name = message.document.file_name
-    with open(file_name, 'wb') as new_file:
-        new_file.write(downloaded_file)
 
-    b = ragDB[message.chat.id].load_file(file_name)
-    bot.delete_message(message.chat.id, sent.message_id)
+    elif (is_valid_attachment(message)):
+        if (message.audio):
+            file_info = bot.get_file(message.audio.file_id)
+            file_name = message.audio.file_name
+        elif (message.document):
+            file_info = bot.get_file(message.document.file_id)
+            file_name = message.document.file_name
+        elif (message.video):
+            file_info = bot.get_file(message.video.file_id)
+            file_name = message.video.file_name
+ 
+        downloaded_file = bot.download_file(file_info.file_path)
+
+    else:
+        downloaded_file = None
+
+
+    if (downloaded_file):
+
+        bot.edit_message_text("Обработка файла...", message.chat.id, sent.message_id)
+
+        with open(file_name, 'wb') as new_file:
+            new_file.write(downloaded_file)
+
+        b = ragDB[message.chat.id].load_file(file_name)
+        os.remove(file_name)
+    else:
+        b = False
+
+    bot.delete_message(message.chat.id, sent.message_id)    
     if b:    
         answer = "Файл загружен успешно"
     else:
-        answer = "Ошибка обработки файла"
+        answer = "Ошибка загрузки файла"
 
     m = bot.send_message(message.chat.id, answer, reply_markup=keyboard)
     prev_messages[message.chat.id] = m    
-
-    os.remove(file_name)
